backend/articles/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
defusedxml==0.7.1
diff-match-patch==20230430
Django==5.0.4
django-categories==1.9.2
django-category==2.0.1
django-ckeditor==6.7.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-import-export==3.3.6
django-js-asset==2.2.0
django-mptt==0.16.0
django-object-tools==2.0.0
django-simple-history==3.5.0
django-taggit==5.0.1
djangorestframework==3.14.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
idna==3.6
inflection==0.5.1
joblib==1.4.0
MarkupPy==1.14
numpy==1.26.4
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pika==1.3.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
threadpoolctl==3.5.0
typing_extensions==4.9.0
unicode-slugify==0.1.5
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.1.0
xlrd==2.0.1
xlwt==1.3.0
